+++ package.json
@@ -13,7 +13,7 @@
   "devDependencies": {
     "@biomejs/biome": "catalog:",
     "lefthook": "catalog:",
-    "turbo": "catalog:",
+    "turbo": "2.1.2",
     "typescript": "catalog:"
   },
   "packageManager": "pnpm@9.10.0",
