--- packages/auth/prisma/schema.prisma
@@ -1,26 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider        = "prisma-client-js"
-  previewFeatures = ["driverAdapters"]
-}
-
-datasource db {
-  provider = "sqlite"
-  url      = env("DATABASE_URL")
-}
-
-model User {
-  id              String    @id
-  username        String    @unique
-  hashed_password String?
-  sessions        Session[]
-}
-
-model Session {
-  id         String   @id
-  expires_at DateTime
-  user_id    String
-  user       User     @relation(fields: [user_id], references: [id])
-}
