+++ packages/auth/tsconfig.json
@@ -25,6 +25,6 @@
     "noUnusedParameters": false,
     "noFallthroughCasesInSwitch": true,
   },
-  "include": ["src/**/*.ts", "test/**/*.ts", "src/index.tsx"],
+  "include": ["src/**/*.ts", "test/**/*.ts", "src/index.ts"],
   "exclude": ["src/gen/**/*.ts"]
 }
