+++ packages/auth/wrangler-test.toml
@@ -1,6 +1,6 @@
 name = "auth"
-main = "./src/index.tsx"
-compatibility_date = "2024-01-01"
+main = "./src/index.ts"
+compatibility_date = "2024-09-09"
 compatibility_flags = ["nodejs_compat"]
