+++ pnpm-workspace.yaml
@@ -3,15 +3,14 @@ packages:
   - packages/*
 catalog:
   "@biomejs/biome": 1.9.0
-  "@cloudflare/vitest-pool-workers": ^0.4.24
+  "@cloudflare/vitest-pool-workers": ^0.5.2
   "@hono/zod-validator": ^0.2.2
   "@types/node": ^20.9.0
   hono: 4.6.1
   wrangler: ^3.76.0
   lefthook: ^1.7.15
   miniflare: ^3.20240806.1
-  turbo: 2.1.1
   typescript: ^5.5.4
   vite: ^5.4.4
-  vitest: 1.5.0
+  vitest: 2.0.5
   zod: ^3.23.8
